tests/e2e/requirements.txt
# End-to-End Test Requirements

# Testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-timeout==2.2.0
pytest-cov==4.1.0

# HTTP clients
requests==2.31.0
websocket-client==1.6.4

# Google Cloud
google-cloud-bigquery==3.13.0
google-cloud-pubsub==2.18.4
google-cloud-monitoring==2.15.1
google-cloud-logging==3.5.0

# Utilities
python-dateutil==2.8.2
pyyaml==6.0.1
